requirements.txt
faker
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
Flask-Seeder
python-dotenv
flask-cors
